move 360 contract test setup to conftest
